+++ runhouse/rns/blob.py
@@ -8,7 +8,7 @@ from runhouse.rh_config import rns_client
 from runhouse.rns.api_utils.utils import generate_uuid
 from runhouse.rns.folders.folder import Folder, folder
 from runhouse.rns.resource import Resource
-from runhouse.rns.utils import _current_cluster, _get_cluster_from
+from runhouse.rns.utils.hardware import _current_cluster, _get_cluster_from
 
 logger = logging.getLogger(__name__)
