+++ runhouse/rns/packages/package.py
@@ -9,7 +9,7 @@ from typing import Dict, Optional, Union
 from runhouse import rh_config
 from runhouse.rns.folders.folder import Folder
 from runhouse.rns.resource import Resource
-from runhouse.rns.utils import _get_cluster_from
+from runhouse.rns.utils.hardware import _get_cluster_from
 
 INSTALL_METHODS = {"local", "reqs", "pip", "conda"}
 
@@ -75,13 +75,6 @@ class Package(Resource):
         install_args = f" {self.install_args}" if self.install_args else ""
         if isinstance(self.install_target, Folder):
             local_path = self.install_target.local_path
-            if not local_path:
-                local_path = "~/" + self.name if self.name else self.install_target.path
-            elif not self.install_target.is_local():
-                # TODO [DG] replace this with empty mount() call to be put in tmp folder by Folder
-                local_path = self.install_target.mount(
-                    path=f"~/{Path(self.install_target.path).stem}"
-                )
 
             if self.install_method == "pip":
                 # TODO [DG] Revisit: Would be nice if we could use -e by default, but importlib on the grpc server
@@ -234,29 +227,6 @@ class Package(Resource):
                     "available for your platform."
                 )
 
-    def to_cluster(
-        self, dest_cluster: Union[str, Dict, "Cluster"], path=None, mount=False
-    ):
-        """Returns a copy of the package on the destination cluster."""
-        if not isinstance(self.install_target, Folder):
-            raise TypeError(
-                "`install_target` must be a Folder in order to copy the package to a cluster."
-            )
-
-        dest_cluster = _get_cluster_from(dest_cluster)
-        if self.install_target.system == dest_cluster:
-            return self
-
-        new_folder = self.install_target.to_cluster(
-            dest_cluster,
-            path=path,
-            mount=mount,
-        )
-        new_folder.system = dest_cluster
-        new_package = copy.copy(self)
-        new_package.install_target = new_folder
-        return new_package
-
     def to(
         self,
         system: Union[str, Dict, "Cluster"],
@@ -264,15 +234,19 @@ class Package(Resource):
         mount: bool = False,
     ):
         """Copy the package onto filesystem or cluster, and return the new Package object."""
-        if isinstance(_get_cluster_from(system), Resource):
-            return self.to_cluster(system, path=path, mount=mount)
-
         if not isinstance(self.install_target, Folder):
             raise TypeError(
                 "`install_target` must be a Folder in order to copy the package to a system."
             )
 
-        new_folder = self.install_target.to(system, path=path)
+        system = _get_cluster_from(system)
+        if self.install_target.system == system:
+            return self
+
+        if isinstance(system, Resource):
+            new_folder = self.install_target.to_cluster(system, path=path, mount=mount)
+        else:  # to fs
+            new_folder = self.install_target.to(system, path=path)
         new_folder.system = system
         new_package = copy.copy(self)
         new_package.install_target = new_folder
